+++ alacritty_terminal/src/grid/mod.rs
@@ -376,6 +376,9 @@ impl<T> Grid<T> {
     pub fn clear_history(&mut self) {
         // Explicitly purge all lines from history.
         self.raw.shrink_lines(self.history_size());
+
+        // Reset display offset.
+        self.display_offset = 0;
     }
 
     /// This is used only for initializing after loading ref-tests.